adk-server/requirements.txt
google-adk
google-adk[eval]
google-cloud-aiplatform
rouge-score 
pandas 
tabulate 
langchain
langchain-mongodb
langchain-openai
pymongo
python-dotenv
litellm